src/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=src/requirements.txt src/requirements/requirements.in
#
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
dj-database-url==2.1.0
    # via -r src/requirements/requirements.in
django==4.2.11
    # via
    #   -r src/requirements/requirements.in
    #   dj-database-url
    #   django-cors-headers
    #   djangorestframework
django-cors-headers==4.3.1
    # via -r src/requirements/requirements.in
djangorestframework==3.15.1
    # via -r src/requirements/requirements.in
idna==3.7
    # via requests
psycopg2-binary==2.9.9
    # via -r src/requirements/requirements.in
python-decouple==3.8
    # via -r src/requirements/requirements.in
requests==2.31.0
    # via -r src/requirements/requirements.in
sqlparse==0.5.0
    # via django
typing-extensions==4.11.0
    # via dj-database-url
urllib3==2.2.1
    # via requests
